#136 - Rename: common 폴더 안에 파일 정리
